refactor: move speed_mining behavior to macro
