+++ src/ares/behaviors/macro/speed_mining.py
